requirements.txt
# Custom Application Dependencies
github3.py==2.0.0
requests==2.26.0
boto3
PyGithub

# Module-Specific Dependencies

# Module: Cloud Service Checker
# No additional dependencies required for this module

# Module: Cloud Security Logging
# No additional dependencies required for this module
logging

# Module: Cloud Vulnerability Scanner
# Add module-specific dependencies for AWS scanning
awscli

# Module: Cloud Compliance Checker
# Add module-specific dependencies for Azure compliance checking
azure-mgmt-resource==18.0.0
azure.identity 
google.cloud

